+++ src/cli/test/parser.test.js
@@ -1,6 +1,7 @@
-const parser = require('./parser');
+const parser = require('../parser');
 
 const namedFunctions = [
+  ' functionName (p1) {',  //class member function
   'function functionName(){',
   'function functionName (arg1, arg2) {',
   'functionName (arg1, arg2, arg3) {',
@@ -21,7 +22,7 @@ const namedFunctions = [
   'functionName({arg1, arg2}) {',
   'functionName({arg1, arg2}) { // inline comment after function',
   'exports.functionName = function(arg1, arg2) {',
-  'export const functionName = (communityIds: Array<string>): Promise<?DBCommunitySettings> => {',
+  'export const functionName = (communityIds: Array<string>): Promise<?DBCommunitySettings> => {'
   // unparsable (without proper squash & parse lexer approach - beyond regx)
   // prettier-ignore
   //'export const isAuthedResolver = (resolver: Function) => async (obj: any, args: any, context: GraphQLContext, info: any) => {'
@@ -41,72 +42,101 @@ const namedFunctions = [
   // ) => {
 ];
 
-const namedFunctionExceptions = [
-  'if (typeof require === "function") {',
-  '// function fname (p1) {',
-  'const mapStateToProps = (state) => ({',
-  'const mapStateToProps = (state): * => ({',
-  'render={({ isLoading }) => primaryAction({',
-  'setActiveChannel = e => this.setState({',
-  "const isSerializedJSON = str => str[0] === '{'"
-];
+
 
 describe('getFunctionName extracts function name from valid function signatures.', () => {
-  it(`Gets function name from a line of code representing a function declaration'`, function () {
+  it.only(`Gets function name from a line of code representing a function declaration'`, function () {
     namedFunctions.forEach(namedFunction => {
+      // console.log('namedFunction', namedFunction)
       const res = parser.getFunctionName(namedFunction);
       expect(res).toBe('functionName');
     });
   });
 
-  it(`Does not get function name from a line of code easily mistaken for a false positive function declaration'`, function () {
-    namedFunctionExceptions.forEach(namedFunction => {
-      expect(parser.getFunctionName(namedFunction)).toBe(undefined);
+  it(`identifyFunction returns true for line of code representing a function declaration`, function () {
+    namedFunctions.forEach(namedFunction => {
+      // console.log('namedFunction', namedFunction)
+      const res = parser.identifyFunction(namedFunction);
+      expect(res).toBe(true);
+    });
+  });
+
+  it(`identifyFunction returns false for code easily mistaken as false positive function declaration`, function () {
+    const invalidNamedFunctions = [
+      'if (typeof require === "function") {',
+      '// function fname (p1) {',
+      'const mapStateToProps = (state) => ({',
+      'const mapStateToProps = (state): * => ({',
+      'render={({ isLoading }) => primaryAction({',
+      'setActiveChannel = e => this.setState({',
+      "const isSerializedJSON = str => str[0] === '{'"
+    ];
+
+    invalidNamedFunctions.forEach(signature => {
+      expect(parser.identifyFunction(signature)).toBe(false);
     });
   });
   /*
- export default function declared inline un-named special case.
- the functionName is effectivly the filename.
- further, if the file is named index, its the parent dir name!
- e.g. 'export default (arg1) => {'
- getFunctionName returns 'default' in this case, and it is 
- up to the caller to apply file name / parent dir.
- see: defaultFunctionName
+    un-named special case for 'export default' and 'module.exports ='.
+    the functionName is effectivly the filename.
+    further, if the file is named index, its the parent dir name!
+    e.g. 'export default (arg1) => {'
+    getFunctionName returns 'default' in this case, and it is 
+    up to the caller to apply file name / parent dir.
+    see: defaultFunctionName
 */
-  it(`returns default for default export unnamed arrow functions`, function () {
+
+
+  //export default cases: TODO -refactor, should all return undefined.
+
+  it(`returns null for default export unnamed arrow functions`, function () {
     const namedFunction = 'export default (arg1) => {';
-    expect(parser.getFunctionName(namedFunction)).toBe('default');
+    expect(parser.getFunctionName(namedFunction)).toBe(null);
   });
 
   it(`returns defaultfunction for default export unnamed functions`, function () {
     const namedFunction = 'export default function (arg1) {';
-    expect(parser.getFunctionName(namedFunction)).toBe('defaultfunction');
+    expect(parser.getFunctionName(namedFunction)).toBe(null);
   });
 
   it(`returns defaultfunction for default export unnamed functions, with no params`, function () {
     const namedFunction = 'export default function () {';
-    expect(parser.getFunctionName(namedFunction)).toBe('defaultfunction');
+    expect(parser.getFunctionName(namedFunction)).toBe(null);
   });
 
   it(`returns defaultfunction for default export unnamed async functions`, function () {
     const namedFunction = 'export default async function (arg1) {';
-    expect(parser.getFunctionName(namedFunction)).toBe('defaultfunction');
+    expect(parser.getFunctionName(namedFunction)).toBe(null);
   });
 
   it(`returns default for default export unnamed async arrow functions`, function () {
     const namedFunction = 'export default async  (arg1) => {';
-    expect(parser.getFunctionName(namedFunction)).toBe('default');
+    expect(parser.getFunctionName(namedFunction)).toBe(null);
   });
 
   it(`returns default for default export unnamed arrow functions, with no params`, function () {
     const namedFunction = 'export default () => {';
-    expect(parser.getFunctionName(namedFunction)).toBe('default');
+    expect(parser.getFunctionName(namedFunction)).toBe(null);
   });
 
-  it(`getDefaultFunctionName returns functionName as name of file`, function () {
+
+
+  //defaultFunctionName 
+
+  it(`getDefaultFunctionName for arrowFunction returns functionName as name of file`, function () {
     const namedFunction = 'default';
     const filePath = 'src/component/TestComponent.js';
-    expect(parser.getDefaultFunctionName(namedFunction, filePath)).toBe(
+    const match = 'module.exports = () => {}'
+    expect(parser.getDefaultFunctionName(match, filePath)).toBe(
+      'TestComponent'
+    );
+  });
+
+  it(`getDefaultFunctionName  for function() returns functionName as name of file`, function () {
+    const namedFunction = 'default';
+    const filePath = 'src/component/TestComponent.js';
+    const match = 'module.exports = function (p1) {}'
+    expect(parser.getDefaultFunctionName(match, filePath)).toBe(
       'TestComponent'
     );
   });
@@ -114,7 +144,8 @@ describe('getFunctionName extracts function name from valid function signatures.
   it(`getDefaultFunctionName returns parent dir name if file named index`, function () {
     const namedFunction = 'default';
     const filePath = 'src/component/TestComponent/index.js';
-    expect(parser.getDefaultFunctionName(namedFunction, filePath)).toBe(
+    const match = 'export default () => {}'
+    expect(parser.getDefaultFunctionName(match, filePath)).toBe(
       'TestComponent'
     );
   });
@@ -163,7 +194,7 @@ describe('addLogging (content, config, filePath)', () => {
     };
     const filePath = 'src/testing/defaultTester.js';
     const res = parser.addLogging(content, config, filePath);
-    expect(res).toEqual(`export default () => { return }`);
+    expect(res).toMatch(/autodlog.log/);
   });
 });
 
@@ -172,7 +203,7 @@ describe('addLogging (content, config, filePath)', () => {
 
   it('injects auto logging into an export default unnamed arrow function', () => {
     const content =
-      `export default () => { return }`
+      `export default () => { }`
     const config = {
       "globPattern": "./**/*.?(js|jsx|ts|tsx)",
       "excludes": "node_modules|\\.test|dlog.js",
@@ -182,7 +213,7 @@ describe('addLogging (content, config, filePath)', () => {
     }
     const filePath = 'src/testing/defaultTester.js'
     const res = parser.addLogging(content, config, filePath)
-    expect(res).toEqual(`export default () => { return }`)
+    expect(res).toMatch(/autodlog.log/);
   })
 
   //edge case dlog + on async library