+++ test/e2e/components/TestComponent.js
@@ -1,11 +1,18 @@
-const dlog = require('./../dlog.js');
+const dlog = require ('./../dlog.js');
 
 exports.namedOnExport = function (p1, p2) {
-  dlog.log({ 'namedOnExport': { p1: p1, p2: p2 } }, { arguments })
-};
+  dlog.log( { 'function' : {p1 : p1, p2 : p2} }, { arguments } )
+ };
 
 export default (p1, p2) => {
-  dlog.log({ 'TestComponent': { p1: p1, p2: p2 } }, { arguments })
+  dlog.log( { 'TestComponent' : {p1 : p1, p2 : p2} }, { defaultExport : 'TestComponent', arguments } )
 
   p3 = p1 + p2;
 };
+
+module.exports = (p1, p2) => {
+  dlog.log( { 'TestComponent' : {p1 : p1, p2 : p2} }, { defaultExport : 'TestComponent', arguments } )
+
+  return
+  //
+}